+++ src/core/building/resources.ts
@@ -1,7 +1,7 @@
 import * as path from "path";
 import { sync as globSync } from "glob";
-import { ResourceKey } from "@s4tk/models/types";
 import type { ValidatedPath } from "./summary";
+import { TGI_REGEX } from "#helpers/file-names";
 
 //#region Constants
 
@@ -11,8 +11,6 @@ export const SUPPORTED_EXTENSIONS = [
   ".stbl.json",
   ".xml",
 ];
-
-export const TGI_REGEX = /(?<t>[0-9a-f]{8}).(?<g>[0-9a-f]{8}).(?<i>[0-9a-f]{16})/i;
 
 //#endregion
 
@@ -52,24 +50,6 @@ export function isSupportedFileType(filepath: string): boolean {
   const filename = path.basename(filepath);
   if (SUPPORTED_EXTENSIONS.some(ext => filename.endsWith(ext))) return true;
   return TGI_REGEX.test(filename);
-}
-
-/**
- * Parses a ResourceKey from a TGI filename, if possible. Returns undefined if
- * the filename does not pass the TGI regex.
- * 
- * @param filename Name of file to parse TGI from
- */
-export function parseKeyFromTgi(filename: string): ResourceKey | undefined {
-  const match = TGI_REGEX.exec(filename);
-  if (match?.groups) {
-    const { t, g, i } = match.groups;
-    return {
-      type: parseInt(t, 16),
-      group: parseInt(g, 16),
-      instance: BigInt("0x" + i),
-    };
-  }
 }
 
 //#endregion